Move fetch-page-data to page-anaylsis/background module
- makes more sense there than util; now will only be called through interface in `page-analysis/background` entry index.js
